+++ CHANGES.md
@@ -1,6 +1,18 @@
 ChangeLog
 =========
 
+0.1.8 (2015-12-XX)
+--------------------
+
+Bug Handling
+************
+-
+
+Features
+********
+- zk-stats-daemon now supports --exclude-bytes, which
+  will skip accounting bytes per path and request type
+
 0.1.7 (2015-12-03)
 --------------------